+++ src/sorting/QuickSort.ts
@@ -2,7 +2,7 @@ import { DSArray } from "src/objects/dsarray";
 import { MessagesObject, NBSP } from "~/engine";
 import { compare, updateDefault } from "~/helpers";
 import { Sorter } from "~/sorting";
-import { BaseSorter, SortMessages } from "~/sorting/sort";
+import { BaseSorter, SortMessages } from "~/sorting/BaseSorter";
 
 export const QuickSortMessages = {
     sort: {
